refactor(dojo): single localDate source
